+++ tests/integration/test_mindful.py
@@ -75,6 +75,17 @@ def test_mindful_api_response3(mock_api_response):
     assert mf.endpoint == "responses"
 
 
+@mock.patch("viadot.sources.mindful.handle_api_response", return_value=MockClass)
+@pytest.mark.connect
+def test_mindful_api_response4(mock_api_response):
+    mf = Mindful(header=header)
+
+    response = mf.get_survey_list()
+
+    assert response.status_code == 200 and isinstance(response.json(), list)
+    assert mf.endpoint == "surveys"
+
+
 @mock.patch("viadot.sources.Mindful._mindful_api_response", return_value=MockClass)
 @pytest.mark.save
 def test_mindful_interactions(mock_connection):
@@ -99,9 +110,23 @@ def test_mindful_responses(mock_connection):
     os.remove("responses.csv")
 
 
+
 @mock.patch("viadot.sources.Mindful._mindful_api_response", return_value=MockClass2)
 @pytest.mark.exception
 def test_file_exception(mock_mindful):
     mf = MindfulToCSV()
     response = mf.run(credentials_mindful=credentials_mindful)
     assert response == None
+
+
+@mock.patch("viadot.sources.Mindful._mindful_api_response", return_value=MockClass)
+@pytest.mark.save
+def test_mindful_surveys(mock_connection):
+    mf = Mindful(header=header)
+    response = mf.get_survey_list()
+    mf.response_to_file(response)
+
+    assert mf.endpoint == "surveys" and isinstance(mf.endpoint, str)
+    assert os.path.exists("surveys.csv")
+    os.remove("surveys.csv")
+